Rework labels to allow for different categories of labels
Summary:
Up until this point, we've really only had labels representing the states within the coroutine. Due to t18447567 (and just as general usefulness), it would be nice to have these labels bucketed into different categories. So, for instance, I would like to have labels that represent states of the coroutines //as well as// labels that allow me to jump into the middle of a loop as part of rewriting a `do`-`while` construct. We could just use the same label category (`label0`, ..., `labeln`) for all labels, but that would be difficult to keep track of and would require additional bookkeeping when generating the `switch` construct at the beginning of the generated state machine.

Instead, I've introduced different types of labels, which will help us differentiate between which labels are being used in the generated state machine.

Reviewed By: ericlippert

Differential Revision: D5124764

fbshipit-source-id: 83538a494759328bc9edf92bc1052f101b355538