Convert a for construct into a while-true-with-if-condition construct
Summary:
This is all that's needed in order to support suspension points within `for` constructs. According to PHP.net, a `for` loop has the structure:

```
for (expr1; expr2; expr3)
    statement
```

and

> Each of the expressions can be empty or contain multiple expressions separated by commas. In expr2, all expressions separated by a comma are evaluated but the result is taken from the last part. expr2 being empty means the loop should be run indefinitely (PHP implicitly considers it as TRUE, like C).

So there are four things that I pull out of the `for` declaration:

  1. Initializer expressions
  2. Control expressions (...For lack of a better name. This is everything in the `expr2` position excluding the last expression.)
  3. The condition expression
  4. End-of-loop expressions

If the condition expression is empty, we instead use `true`.

Then, we write a `while` construct with the following template:

  initializer_exprs
  while (true) {
    controls_exprs_except_last
    if (!condition_expr) {
     break;
    }
    body
    end_of_loop_exprs
  }

Since we already desugar `if` and `while` statements, this will desugar as well.

Reviewed By: ericlippert

Differential Revision: D5076165

fbshipit-source-id: fea9e90655a13537b08440b1d0e0e41124dc8a0b