Desugar suspends used in 'foreach statements' in the generated state machine
Summary:
`foreach` statements now permit suspensions in collection positions.

`foreach` also supports an `await` syntax, which is something like:

  public async function singleSuspend(
    AsyncIterator<int> $asyncCollection
  ): Awaitable<void> {
    foreach ($asyncCollection await as $item) {
    }
  }

This diff does not introduce an analogous usage for coroutines. We still need to discuss whether or not we want to allow users to manipulate coroutine invocations that have not been suspended (e.g., `$x = someCoroutine();`) before this would be meaningful.

Reviewed By: ericlippert

Differential Revision: D5115380

fbshipit-source-id: 33bec5c8286b003fe6fafe94556a7acb05b5df1a