Convert a do-while construct into a while-true-with-if-condition construct
Summary:
Previously, we directly lowered

  do {
    BODY STATEMENTS
  } while (suspend someCoroutine());

into

    do {
      BODY STATEMENTS
      SUSPEND REWRITING...
      $closure->coroutineResultData1 = $coroutineData;
      ...
    } while ($closure->coroutineResultData1);

This is bad, because if `BODY STATEMENTS` contains a `continue`, then we will erroneously skip over the `SUSPEND REWRITING`.

The fix is to lower a `do`/`while` construct into a while-with-goto-construct before we do the suspend rewriting. So, (before we do the suspend rewriting), we lower

  do {
    BODY STATEMENTS
  } while (CONDITION);

into

  goto loop_label_#;
  while (true) {
    if (!CONDITION) {
      break;
    }

    loop_label_#;
    BODY STATEMENTS
  }

Voilà, no more `do`-`while`s.

Reviewed By: ericlippert

Differential Revision: D5131236

fbshipit-source-id: 8475ec9626aee6189e20b3d6adcbc62d468e34c6