Make XHP splitting behavior more conservative
Summary:
Previously, we were adding splits between XHPExpressions and XHPBody tokens. This causes problems when an XHPBody token needs to have no whitespace between it and an adjacent XHPExpression, like in quoted fbt expressions, which can be broken by the added splits.

This diff ensures that we add splits only when we can do so without introducing significant whitespace. Since whitespace is significant when it is adjacent to an XHPBody token, we can safely add splits anywhere *not* adjacent to an XHPBody token, as well as adjacent to an XHPBody token that already has whitespace on the side where we are adding the split.

Differential Revision: D4984152

fbshipit-source-id: ec1816ad0dffc1ed662df924856aa39c60e96feb