Compare closure classes dynamically
Summary:
Closure classes shouldn't be compared by name, but by the way they're used. So we change
semdiff to schedule them for comparison when we hit CreateCl instructions on the two
sides when comparing other bodies. We then look up which classes are referred to and compare
them (which may schedule further class pairs to be compared).

This is a half-way house diff, in that all classes should really be compared like this, scheduled
by the DefCls instructions in .main and other top-level functions. But I'm holding off on
that change until hack agrees with hhvm about emitting those DefCls instructions.

Reviewed By: hubyrod

Differential Revision: D5154814

fbshipit-source-id: 3ec396d02f105141ec3a49c751190ac68b5944a9