report server exception to persistent client
Summary:
If the server encounters an exception, it abruptly closes the connection to the persistent client.

This diff lets us at least report the exception and its stacktrace to the client before closing. Later in the stack, we'll log those exceptions+stacktraces to scuba and to the nuclide log.

That will give us the data we need to start zeroing in on the causes of exceptions that close the persistent connection, and fixing the important causes.

Reviewed By: arxanas

Differential Revision: D5143657

fbshipit-source-id: 5efc6bccccdfbb10305432df8c139a6611ff44c1