Fix a bug in unwind-itanium.h
Summary:
We've been getting away with this, because we only really use it
to identify InvalidSetMException and TVCoercionException.

Because of the way we throw/catch them, they're never dependent
exceptions, and I guess false positives are unlikely.

With some other changes I'm working on, I'm seeing crashes because
we're getting garbage type-infos here.

Reviewed By: ricklavoie

Differential Revision: D5150176

fbshipit-source-id: 663388fe6a87208169c38b9cb48a5d98f41230ee