Catch unexpected exceptions at the php boundary
Summary:
If a builtin throws a std::exception, or something derived from
std::exception (but not ExtendedException) its mostly handled
correctly, but results in an error message that contains the mangled
name of the exception, and no backtrace.

This ensures that any such exceptions are caught and converted to
FatalErrorExceptions when we start unwinding, which will give them a
proper backtrace.

It does mean that we can only throw req::root<Object>s, and
std::exceptions; anything else will crash.

Reviewed By: paulbiss

Differential Revision: D5145802

fbshipit-source-id: 6448db798084dd3d450d41a4b3871476ba873a87