Move ext_watchman back to stable state
Summary:
D4816886 was supposed to improve code readability and fix a subtle bug, but in practice we have a workaround for the bug in PHP and test pass rates have massively regressed.

This diff undoes the code-refactoring but keeps in subsequent fixes found during this exercise including: D5112492, D5112492, and adding `s_allClients.clear()` to `moduleShutdown()`.

Reviewed By: markw65

Differential Revision: D5155486

fbshipit-source-id: 6caed68e042a527086dbf1f4ccefea5cce67a75d