Resizable buffers for RetranslateAll
Summary:
Previously per-function buffers for RetranslateAll had been fixed size,
necessitating a very large initial allocation to account for a few large
functions present in www. This was creating a great deal of memory pressure
during warmup and pegging process virtual memory (jemalloc would madvice away
the allocations but not release the pages). This diff introduces the notion of a
resizable DataBlock, which uses malloc/realloc/free to manage an internal
buffer. The mode is only supported for DataBlocks with virtual base addresses as
the underlying storage may need to be moved to a new base.

To avoid thrashing the allocator creating individual buffers for each DataBlock
of each function being optimized this approach uses an inital fixed size
contiguous buffer as the initial storage for all local DataBlocks. If blocks
outgrow their initial buffers the create a dynamic buffer and memcpy the initial
buffer over. The dynamic buffer is then resized using realloc and managed by
a std::unique_ptr member of DataBlock.

This diff also removes the persistent thread local TC caches and debug mode
mprotect guards. The guards were difficult to make work with the new resizable
buffers, and the thread local mode was never used outside of testing.

Reviewed By: swtaarrs, ottoni

Differential Revision: D5141564

fbshipit-source-id: 2b37970f2beaf248df10b4e182b21fb20e26a688