Fix ServiceData::exportCounterByKey() for callback counters
Summary:
I missed this in D5117272. For now, if a callback doesn't appear in
the regular counter map, get all the callback counters and check the result for
the key we want. If we ever have enough callback counters for this to be too
slow, we can change to a map-based solution.

Reviewed By: intellectuallyswole

Differential Revision: D5166202

fbshipit-source-id: cdbbe95e181a7ea08b40adb039621fc24b993db6