ClientMessageQueue reports exceptions+stacktraces
Summary:
This diff cleans up the way ClientMessageQueue reports errors...

(1) We want to always capture the exception and its stacktrace. That will let us, later in the diff, report it to scuba+nuclide. This diff captures exceptions at the point they're raised in the daemon and marshals them over to the queue reader.

(2) Since our abstract data type has to include exceptions with their stack traces, it was easier just to use this to also represent End_of_file exceptions, rather than No_more_messages.

(3) I tightened up the categorization of which failures are recoverable and which ones aren't. Basically, it's recoverable if it won't affect the way we read future JsonRPC messages. (That means for instance that a missing Content-Length header is unrecoverable).

Reviewed By: arxanas

Differential Revision: D5143658

fbshipit-source-id: 8c89551a4da2328b38fb70509699e2cf6c7a5104