Fix a bug in subtypeOfData
Summary:
With the new test case, on the first analysis pass it looks like
there's a path through the catch block, and we get a possibly empty
array (also it ends up as a ArrLikeMapN). On the second pass it can
eliminate that path, and produces a more refined type; but due to a
bug, hhbbc didn't recognize it as more refined.

Reviewed By: ricklavoie

Differential Revision: D5189183

fbshipit-source-id: f9732947b39af4684e12dfcb2e2f606bc6640c9a