Deal with AttrNoSerialize in unserialize path
Summary:
We correctly would not serialize a property marked with AttrNoSerialize, however
the unserialize path did not take this into account. This meant that a
hand-crafted serialized string could be unserialized and overwrite such
properties. This meant that something like the memoize caches could be changed,
breaking various invariants. Check for AttrNoSerialize in the unserialize path
and ignore such data.

Reviewed By: markw65

Differential Revision: D5173899

fbshipit-source-id: f4514cf3aadcd3694ab38a09855c16f557751c81