Move generated state machine out of helper method, into lambda
Summary: We can avoid generating an extra helper method, and thereby polluting the name space of the enclosing code, by generating the state machine method as a lambda. That way all the code stays inside the method it came from. Also, this mechanism is more general, allowing us to use the same sort of transformation on a function, a method, or a lambda coroutine.

Reviewed By: michaeltingley

Differential Revision: D5143038

fbshipit-source-id: e2fa8162e8ab1f72e7d93bca4a81a52e7b3ee5e2