Setting up testing infrastructure
Summary:
Currently, tests where a functionality such as autocomplete or identifying a symbol needs to be run at a specific location in a target file mark said location in one of two ways. Either the location is specified in (line, character) format in a separate `*.flags` file or a location in the target file is marked using a keyword such as "AUTO332".

See `/hphp/hack/test/autocomplete` or `/hphp/hack/test/identify_symbol` for examples.

We wish to move away from this format by specifying the target (line, character) directly in the test file as a comment near the top of the file. The actual FFP autocomplete is still a work in progress so the current test output is just a temporary placeholder.

Running an ffp_autocomplete test on a file will find the first comment of the form:
/* [test type] [param1] [param2] */
and output this information to standard out.

Reviewed By: arxanas

Differential Revision: D5159967

fbshipit-source-id: 1f1831fbdea0f431289970ac4c770103fbb4d90f