Lambda analyzer for coroutine rewriting
Summary:
The coroutine rewriter will need to be able to determine the outer variables of a lambda given its position in a parse tree. This code works that out and returns a sorted list of variables declared outside a lambda which the lambda actually uses.

The logic might not be quite right yet for old-style anonymous functions, which can declare in a using clause what their captured outer variables are. I'll work that out in a later diff; right now I want to get lambdas sorted out.

Reviewed By: michaeltingley

Differential Revision: D5104379

fbshipit-source-id: 75c4b5b1c4cd0bf951c568315b01fcb2f5d4a34c