Fix unwinder crashes
Summary:
If we don't find the typeinfo in typeInfoMap, we return
_UA_HANDLER_FOUND from the search phase, promising to handle the
exception in the cleanup phase. But if another thread updates the
table before we get back to the cleanup phase, we might decide not to
handle it, which will lead to an assertion.

This adds a cache, which may help performance, but guarantees that the
cleanup phase sees the same thing the search phase does.

Reviewed By: ricklavoie

Differential Revision: D5194534

fbshipit-source-id: c969569fbfba309f7e3faa742c8bddc0d7f57937