Copy-in-copy-out locals for coroutines
Summary:
The original idea was to rewrite all usages of locals like $x into $closure->x in a coroutine.

This presents several problems. For example:

* Many syntaxes do not rewrite cleanly. For example `"$x->y"` cannot be rewritten to `"$closure->x->y"` without changing the meaning; it would have to be `"{$closure->x->y}"`.
* Capture semantics for non-coroutine lambdas end up capturing the wrong thing. We'd have to introduce an extra layer of closure around each non-coroutine lambda.
* And so on.

It is much easier to simply copy in the set of locals when the coroutine resumes, and copy them back to the closure when it returns, suspends or throws.

This has some costs:

* Coroutines that are frequently resumed but use few locals means we impose a high cost per resumption.  We could write an optimizer that copied only the locals that were used.
* The closure does not have up-to-date values for the captured locals

I note also that there are improvements we could make:

* We could skip hoisting locals which are not used across suspensions.

Reviewed By: michaeltingley

Differential Revision: D5186473

fbshipit-source-id: fcccdb4c09f06acd972181bbdc552b894fee4400