Omit trailing trivia from at-char output
Summary:
Currently, trailing line comments are included in the output of --at-char, causing the cursor to jump from (for instance) a semicolon that was typed to the end of a trailing single-line comment following it. This happens because we treat chunks as indivisible, and a semicolon and trailing line comment end up in the same chunk because we never split between them (nor should we, since doing so would allow the trailing comment to be placed on a new line, making it appear to annotate the following line instead of the preceding line).

To deal with this, this diff adds some metadata to chunks regarding the tokens contained in the chunk. This metadata associates the location of the token in the original source with its location within the chunk, giving us the ability to print ranges that fall at token boundaries within a chunk.

Reviewed By: arxanas

Differential Revision: D5161337

fbshipit-source-id: ac974d817aa1234c469ca0fde5fc0099a7b90b5a