revert "Persistent_soft"
Summary:
I'd earlier changed "Persistent" into "Persistent_hard / Persistent_soft" to indicate whether an attempt to gain the persistent connection should aggressively kick off any other existing persistent connection, or just back off.

D5015027

But as I got to the end I discovered some flaws: (1) there was a fairly clunky UX where it popped up a notification inviting the user to "retry" to gain back control of the persistent connection for their window... we could have done away with this UX with extra work in Nuclide but not in VSCode; (2) Each time it popped up this dialog, it counted as a crash, and the editor gives up after 5 crashes in 3 minutes... again we could have done away with this in Nuclide by extending LSP, but not in VSCode.

I have a better plan up my sleeve... (at the tip of this stack).

Reviewed By: arxanas

Differential Revision: D5176211

fbshipit-source-id: 6e68d2bda54338ce1d4f18faddbbde096b8f10ff