move LSP state more local
Summary:
This diff moves a few things out of global mutable state fields, and into immutable "env" of the current mode (e.g. Main_loop, Pre_init).

It's a picky change that affects loads of lines in trivial ways. Sorry about that.

I'm making this change because it feels cleaner, more in keeping with the way In_init already works, and more in keeping with the way the rest of Hack is authored. (with "env" variables being passed around).

I'm also making this change because later in the stack I want to add some extra state fields specific to the Main_loop state. And the only good place to put them is inside a "Main_env". And if I'm going to be putting some stuff in Main_env, then I might as well put all of it.

Reviewed By: arxanas

Differential Revision: D5176235

fbshipit-source-id: 9ca556e1dca52320c9323f08cd3098c63a98a20c