add a new LSP state "Lost_server"
Summary:
If we had a persistent connection, but then a rival window grabs it, we will enter this new state "Lost_server".

(Later in the stack, if we're in the "Lost_server" state and the client requests some action like editing or typing or hover, then we'll automatically get the persistent connection back).

Reviewed By: arxanas

Differential Revision: D5176580

fbshipit-source-id: 8115513f54afe2c27c53ddf1a5a24dfb0821f9e6