upon losing server, flush diagnostics
Summary:
When another Nuclide window gains control of the persistent connection, then our window can't be showing diagnostics anymore. (they will become stale).

This diff keeps track of which files we've reported a non-empty set of diagnostics on so far, so that in the event of server loss, we can tell the client to get rid of them all.

Note that if the LSP connection actually *closes* then it's the responsibility of the client to flush diagnostics. We have to do it here only because in the event of server-being-stolen then we keep the LSP connection alive.

Future work: I thought it would be nice, if there had been errors which we got rid of, to replace them with just a single warning message that says "Your errors are in a different castle". But that can be left for later, and only if people really are confused about it.

Reviewed By: arxanas

Differential Revision: D5176716

fbshipit-source-id: 690124b8f2f418e252e2c0e3a36a80ff2e75f3d3