Delete symbol-by-id
Summary:
There used to be a Nuclide idea to do "symbol-by-id", which would tie up all the company's symbols (across languages) with wiki/dex, and Nuclide would have a "Context view" menu which let you see it.

That never really took off, and is being removed from Nuclide. So I'm removing it from Hack as well.

Reviewed By: dabek

Differential Revision: D5201655

fbshipit-source-id: e567689fb99ce125c9a8aa538be5de4fad101be0