Deal with FPassL, remove IterBreak exception
Summary:
Removed IterBreak exception altogether - uses of IterBreak with which we
can't cope now fail gracefully instead of causing a gratuitous exception.

Dealt with FPassL instruction. There are some examples, notably relating to the
compilation of the pipe operator, in which this instruction is used with an
unnamed local variable. But the "instruction behaves like VGetL" if the relevant
parameter is passed by reference, and (1) we don't currently deal with VGetL on
unnamed locals, because we can't track what use is made of the reference once
we've lost sight of it in the stack, and (2) semdiff doesn't know if the parameter
*is* pass by reference or not, so we should conservatively assume it could be. So
I've implemented a rather special-purpose pattern: iff the FPassL is immediately
followed by a sequence of other FPassX instructions, followed by a call instruction
that will consume all the parameters pushed in that sequence, then we treat the
FPassL instruction as if it read and then wrote the local variable, which is a
conservative approximation of what might really happen if it's passed by reference
and mutated by the called function.
This is pretty weak (it gives up when presented with nested calls with unnamed locals passed
this way, for example), but it suffices for the pipe-like examples.

Also added translation of Silence instruction to hhbc_hhas, as its omission made
one of the IterBreak tests fall over when trying to print the diff.

Reviewed By: andrewjkennedy

Differential Revision: D5201034

fbshipit-source-id: 4fe36dbacdde185d9d7419908c687c3e4e97bcd8