Watchman event watcher tracks repo state
Summary:
Watchman event watcher tracks repo state

Eventually, the state watcher is going to listen in on a socket
for connections where clients can ask if the repo is mid-update,
and if it is get notified when the update ends. This will avoid
various races that could wedge a server into eternally waiting
for a state leave.

Reviewed By: dlreeves

Differential Revision: D5128659

fbshipit-source-id: acf28045af3c6af2da3905217b2dd826bb6e95d5