use immutable queue for file-edits
Summary:
While waiting for the server to ready, we have to queue up file-edits so we can apply them all once it is ready.

Previously we did this with a list (which we had to reverse before applying).

This diff switches it over to an immutable queue, which is a better fit.

Reviewed By: nmote

Differential Revision: D5195272

fbshipit-source-id: 5bcf3d85aafc622338262f3f1cac3f4483ba1e14