Pull in .hhi files when performing the type checking test for lowered coroutines
Summary:
Prior to this, we pulled in `ai_builtins.ml`'s definitions for a few builtin constructs. While this file is used for type checking, it is really the .hhi files that are authoritative in terms of defining the standard library. As we build out the coroutines standard library, it will be not only tedious but also dangerous to attempt to keep `ai_builtins.ml` in sync with the committed .hhi file changes.

Instead, `coroutine.ml` now slurps up all of the .hhi files it can find in the provided directory (specified with `--hhi-dir`), and adds them the parsing/naming/declaration/type checking that it performs during the test. I've updated the `TARGETS` files to ensure that the .hhi files will be available in the testing environment, and that the testing environment targets the correct path when slurping up the files.

Finally, I've reverted the coroutines changes to `ai_builtins.ml`, in order to discourage others from using this file for type checking purposes.

Differential Revision: D5162249

fbshipit-source-id: a41d11de05171dd5fdd854dc453d048088e129c7