Make arguments to the coroutine method untyped when received in the closure
Summary:
On their own, keeping ordinary types around isn't a problem.

The problem comes when the arguments don't have ordinary types, but instead have generic types. In this case, we would have a problem passing through the argument to the closure. In the coroutine function, we would have, for example, `T $foo`. The closure would receive this as `T $foo`. Unfortunately, without extra work by the compiler, the closure is not generic, and so Hack would complain that it has no idea what `T` is referring to. Instead, we can just receive the parameters are type-less, and the problem goes away.

Differential Revision: D5170729

fbshipit-source-id: eaffc30fd4e36e89c789f169278181e288c8e9df