Specialize the type of the closure's received state machine lambda to correctly identify the type of the original coroutine funciton
Summary: The generated state machine lambda correctly returns a value of type `CoroutineResult<ReturnType>` (where `ReturnType` is the return type annotation of the original function). However, in the generated closure, we were receiving this the generated state machine lambda as having a return type of `CoroutineResult<mixed>`. This is a small diff to specialize that annotation.

Reviewed By: ericlippert

Differential Revision: D5161166

fbshipit-source-id: 9f85e39bdd58f69b4c5b4ca62e357f6f1d60a2a6