Lower suspend invocations by passing the closure through to the invoked coroutine function
Summary:
Let's say we suspend a function invocation.

  public coroutine function suspendExpression(): void {
    suspend someCoroutine();
  }

Previously, when we lowered this "suspend invocation", we added the decorations to track the suspend state, handle possible exceptions on resumption, etc. However, the actual coroutine invocation would remain unchanged. The invoked coroutine function, however, requires the current `$closure` to be passed in, as //that// coroutine's completion continuation. This diff implements that behavior. Particularly:

```lang=diff
(
  Class_suspendExpression_GeneratedClosure $closure,
  mixed $coroutineData,
  ?Exception $exception,
): CoroutineResult<CoroutineUnit> ==> {
  // ... many lines omitted ...
- $coroutineResult = someCoroutine();
+ $coroutineResult = someCoroutine($closure);
  // ... many lines omitted ...
}
```

It's worth catching that the type checker tests would have caught this, but so far we have only been `suspend`ing on nonexistent functions, and so it didn't have anything to type check. I've added a test that exercises suspending on a known function. With the changes in this diff, the test now passes.

Reviewed By: ericlippert

Differential Revision: D5161232

fbshipit-source-id: 84c2abb1b553f995253600f99d25283eba52c69a