Start refactorings to lower coroutine lambdas
Summary:
I am embarking on a series of changes to facilitate lowering of coroutine lambdas.

In this diff: today a coroutine method body is lowered, which produces a lowered body and a set of locals to be captured in the closure.

This is problematic, because a coroutine method body that contains coroutine lambdas may have to generate multiple closure classes.

I've started by refactoring so that coroutine method body lowering returns the lowered body *and the closure class*.

The next refactoring will be to return several closure classes.

Reviewed By: michaeltingley

Differential Revision: D5204400

fbshipit-source-id: f747f7be7af53b92257c081eb73a44c8e8e4dc10