Hack codegen: omit guards in Memoize code when function definitely does not return null
Summary:
HHVM uses a special "guard" property in the Memoize code, but doesn't need this if the return value can never be null.
For now, enable this optimization for `async` methods and functions. In future, introduce a compiler flag to optimize
for non-nullable return types too, as is the case in HHVM.

Reviewed By: oulgen

Differential Revision: D5208878

fbshipit-source-id: ffbcb08de621d8bc8d7377c9d6d0c186b7ea16ef