Compact/cached backtraces
Summary:
Introduce a compact format for storing backtraces on exceptions (provided that
the trace only uses the default options and does not capture arguments). Also
add an LRU cache for storing expanded backtraces and use it for compatible
calls to debug_backtrace in addition to exceptions.

In order to cover more cases I needed to change the behavior of the existing
Eval. EnableArgsInBacktraces to drop the args member of the frame entirely so
that it was identical to `debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS)`

Also introduce a flag to enable/disable this new exception format, default it to
enabled.

Reviewed By: ottoni

Differential Revision: D5144736

fbshipit-source-id: fc1dc9e5771e35c400de58e58886b84c84936385