Don't allow TInitCell as HHBBC specialized array key types
Summary:
Improve HHBBC's array-key handling and add additional invariants. All
specialized array types can't have key types more general than TArrKey (so no
TInitCell). Furthermore, keyset specialized types should never have value types
more general than TArrKey. Finally, keyset specialized types should always have
the same type for the key and value.

Having InitCell appear as a specialized array's key type is problematic because
it can cause us to infer a worse type for certain operations when an array gets
refined. Say, initially we think an array is just a TArr. We don't have any
specialized information so we assume that the array's keys as TArrKey. Later on,
we refine the TArr to a specialized array type, which happens to think the key
is TInitCell. We now perform the same operation and use the specialized
information, and get TInitCell as the key type. The type has gotten worse.

Update the array-key handling to maintain these variants. Add an additional bool
to the ArrKey type (returned from the disect functions) which indicates whether
the key coercion can potentially throw. Use these boolean rather than relying on
key coercions which might throw having their effective type computed as
InitCell. This might improve performance when HackArrCompatNotices is turned on
because we can use the more specific type information without losing the factor
that it can raise a warning.

As a result of this change, the logic to turn a chain of AddElemCs into a static
array needs to pay attention to whether the AddElemC can throw. Don't optimize
away the AddElemC if it can throw.

These improvements fixed an existing bug where we could infer than an array was
non-empty when it could be empty. This happened when the key could be an object
or other strange key. We would assume the key was ArrKey and thus not strange.

Reviewed By: markw65

Differential Revision: D5213726

fbshipit-source-id: f31267c9120107de32bf44b584a5e9e4c112e3a7