Use the Watchman Event Watcher client in the informant
Summary:
Use the Watchman Event Watcher client in the informant.

This is the ultimate payoff in this stack of diffs. In short, we are now able to avoid needlessly slow server inits by delaying init while the repo is mid-flight of an "hg update" until it is complete - at which point init will load the saved state for the target revision.

Reviewed By: dlreeves

Differential Revision: D5163495

fbshipit-source-id: 42250d888738bb7202748b98e87cc87d9a4b0a8d