First pass at an hhas fuzzer
Summary:
This is a first pass at a fuzzer for hhas. It takes a .hhas file as input, and performs semi-random mutations to it to hopefully generate hhas that is broken in nontrivial ways, primarily so we can add more features to the verifier.

This is still very much in development, so any and all suggestions for improvements, new features or changes in general are very welcome.

Reviewed By: nbenton

Differential Revision: D5210890

fbshipit-source-id: dfc606c8aaf6dc9e265388fc3a87ee0972e63032