Light hphp/.clang-format file
Summary:
A clang-format file which incorporates the generally accepted hhvm code style guidelines
(left-aligned pointers, two-space indentation), and tries to be as permissive as possible about everything else.

If you want to use it to format a file, you can `clang-format -i FILENAME`.

This is not incorporated in any linters, and its use is entirely voluntary.

If you dislike the resulting formatting, you can either not use it, or change the rules to suit your needs.

Reviewed By: swtaarrs

Differential Revision: D5205163

fbshipit-source-id: cf4bae685d787ec963eb771100413b22a4289c16