Replace int "rule values" with bool
Summary:
Currently, solve states associate rules (which govern some set of splits) with int values, where a value of 0 means the rule is not broken in the solve state (and thus its associated splits will not be printed as newlines), and 1 means the rule is broken (and its splits will become line breaks in the printed solution).

Originally, the need for int values comes from dartfmt, which had more than two values, representing different possible breakage configurations for the rule's associated splits. We don't need or want to do this, so we can simplify things by constraining the type to bool and removing the confusing name "value".

Depends on D5189148

Reviewed By: arxanas

Differential Revision: D5189162

fbshipit-source-id: 4a42ee62c17a2f5324d4a174260b286e7e50fe28