Don't treat State_queue as immutable
Summary:
Minor refactor for clarity.

We currently return references to a State_queue.t as if it were immutable and we needed to return a reference to an updated version after calling a mutation function. It is mutable, though, so code that passes references around as if it's immutable is misleading.

Reviewed By: arxanas

Differential Revision: D5189238

fbshipit-source-id: 4d0c0d57b6472f8fe0f492919a9fc5be0f6427b7