Speed up type-scanner generation using concurrency
Summary:
Greatly speed up type-scanner generation by parsing each compiliation-unit
concurrently in different threads. This works (for now) because debug
information for each compiliation-unit does not refer to anything outside of
that compiliation unit, so they can be easily processed separately. The initial
scan of all the compiliation-units to find object names and offsets during
parsing dominated the time to generate the scanners.

The second largest amount of time was spent in the actual generator, iterating
over the list of found objects, categorizing them, and building up some
ancilliary data structures. This too can be done in multiple threads, taking
advantage of the fact that the parser state is already now divided up into
discrete blocks.

Reviewed By: markw65

Differential Revision: D5226843

fbshipit-source-id: 06c4e8ec2118d5858498ffd78a2b30c37e5bdce6