Allow indexing into optional shape fields when the expression is part of a null coalescence
Summary:
Many users have complained that I am taking away `??` as a useful part of shapes with our migration to optional/unknown fields. This diff supports that behavior.

= Problem statement

An example is worth a thousand words. Let's say we have this shape definition:

  type ShapeWithOptionalField = shape(?'a' => int);

You can `Shapes::idx` into it:

  hh_show(Shapes::idx($shape, 'a'));
  // ?^(int)

This returns an `int`. However, you cannot not directly index into it:

  $shape['a'];
  // Invalid index operation: 'outer_optional_field' is marked as an optional shape field. It may not be present in the shape. Use Shapes::idx instead. (Typing[4165])

A common pattern in the codebase was this:

  $shape['a'] ?? 12;

Which means, index into this shape using the `'a'` field, and if the field is absent or `null`, use 12 instead. However, my shapes changes simply spat out the following error message:

  Invalid index operation: 'outer_optional_field' is marked as an optional shape field. It may not be present in the shape. Use Shapes::idx instead. (Typing[4165])

This is because the indexing operation was context-insensitive, and so I didn't know that the null coalesce operator was there to "protect" the indexing operation. (I verified that HHVM does not crash when doing this and the `'a'` field is absent from the shape.)

= This diff's fix

This diff institutes context sensitivity for optional shape field indexing. The test cases are really better than my own explanation. These are the result **after** the diff's changes are introduced.

```
type InnerShape = shape(?'inner_optional_field' => int);
type OuterShape = shape(?'outer_optional_field' => InnerShape);
```
== Case 1
```
function foo(InnerShape $shape): void {
  $shape['inner_optional_field'] ?? 12;
}
```

  No errors.

== Case 2
```
function foo(OuterShape $shape): void {
  $shape['outer_optional_field']['inner_optional_field'] ?? 12;
}
```

  No errors.

== Case 3

```
function foo(bool $condition, OuterShape $shape): void {
  ($condition ? $shape['outer_optional_field'] : $shape['outer_optional_field'])['inner_optional_field'] ?? 12;
}
```
```
Invalid index operation: 'outer_opt' is marked as an optional shape field. It may not be present in the shape. Use Shapes::idx instead. (Typing[4165])
```

Note that HHVM in fact crashes if you try to do something like this, so it's important that the type checker prevents this case!

 ---

Reviewed By: dlreeves

Differential Revision: D5202997

fbshipit-source-id: da29b75aa234bfced53a6d553950163546bbde13